Assignment4/requirements.txt
asgiref==3.5.2
boto3==1.26.31
botocore==1.29.31
Django==3.1.6
djangorestframework==3.12.2
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
